inicio jpa(dependências e entidade de produto)/PARCIAL

+++ src/main/java/com/github/felipeanjosunifil/projeto_mvc/model/service/ProdutoServiceImp.java
@@ -1,6 +1,6 @@
 package com.github.felipeanjosunifil.projeto_mvc.model.service;
 
-import com.github.felipeanjosunifil.projeto_mvc.model.Produto;
+import com.github.felipeanjosunifil.projeto_mvc.model.entity.Produto;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -9,7 +9,7 @@ import java.util.List;
 @Service
 public class ProdutoServiceImp implements ProdutoService {
 
-    private int id = 0;
+    private Long id = 0L;
     private List<Produto> produtos = new ArrayList<>();
 
     @Override
